Переписала запрос для получения товаров с CreateAsyncThunk на RTK query
