+++ package.json
@@ -3,6 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@pbe/react-yandex-maps": "^1.2.5",
+    "@radix-ui/react-checkbox": "^1.3.3",
     "@reduxjs/toolkit": "^2.9.0",
     "@uidotdev/usehooks": "^2.4.1",
     "axios": "^1.12.2",
@@ -16,7 +18,6 @@
     "qs": "^6.14.0",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
-    "react-hot-toast": "^2.6.0",
     "react-loading-skeleton": "^3.5.0",
     "react-redux": "^9.2.0",
     "react-router-dom": "^6.30.1",
