+++ src/hooks/useSetQueryParams.ts
@@ -1,5 +1,6 @@
+import { useGetAllProductsQuery } from "@/api/product/productApi";
 import { FilterSliceState, PaginationType, SortItem } from "@/redux/filterSlice";
-import { fetchProducts } from "@/redux/productsSlice";
+// import { fetchProducts } from "@/redux/productsSlice";
 import { useAppDispatch } from "@/redux/store";
 import qs from 'qs';
 import { RefObject, useEffect } from "react";
@@ -35,20 +36,32 @@ function useSetQueryParams( sortTypeValue: SortItem, categoryId: number, paginat
         
     }, [sortTypeValue, categoryId, pagination.currentPage, pagination.limit, selected]);
 
-    useEffect(() => {
-        dispatch(
-            fetchProducts({
-                sortTypeValue, 
-                categoryId, 
-                pagination: {
-                    currentPage: pagination.currentPage,
-                    limit: pagination.limit
-                },
-                selected
-            })
-        );
+    const { data, error, isLoading } = useGetAllProductsQuery({
+        sortTypeValue, 
+        categoryId, 
+        pagination: {
+            currentPage: pagination.currentPage,
+            limit: pagination.limit
+        },
+        selected
+    });
+
+    return { data, error, isLoading };
+
+    // useEffect(() => {
+    //     dispatch(
+    //         fetchProducts({
+    //             sortTypeValue, 
+    //             categoryId, 
+    //             pagination: {
+    //                 currentPage: pagination.currentPage,
+    //                 limit: pagination.limit
+    //             },
+    //             selected
+    //         })
+    //     );
         
-    }, [sortTypeValue, categoryId, pagination.currentPage, pagination.limit, selected]);
+    // }, [sortTypeValue, categoryId, pagination.currentPage, pagination.limit, selected]);
 }
 
 export default useSetQueryParams;
